Add the extra steps taken at Damcyan.
This required adding additional data to support allocating a single
variable to multiple paths on multiple maps. Additional data will
eventually be required to support steps at Fabul and the Dwarf Castle,
but it should be relatively simple to add the necessary data.

I'm not entirely happy with the way the extra steps at Damcyan display
in text form, but it's probably fine.

Signed-off-by: Jason Lynch <jason@calindora.com>