Add Antlion B2F.
This required adding support for including/excluding sequences of maps
depending on the value of a variable in order to support alternate paths
determined by certain variables. The code appears to be both working and
general enough to apply to all scenarios I've thought of. It remains to
be seen if changes will need to be made later nonetheless.

Signed-off-by: Jason Lynch <jason@calindora.com>