Add the ability to display custom routes.
This functionality is largely intended (for the moment) as a way to
deliver customized routes to people if they breka their step route
during a run.

For now, this means including a pre-compiled copy of rosa. Ideally, this
would actually be built as part of the deployment process, but that's a
lot of extra complexity for minimal benefit (especially as this code
base is probably a dead end, ultimately).

Signed-off-by: Jason Lynch <jason@calindora.com>