Add the initial version of the no64-rosa-marathon-recovery-fusoya routes.
These routes are designed to allow the player to recover a step route by
saving and resetting immediately upon arriving at the moon the first
time. The seed finder tool has been updated to support this route. The
routes are not quite fully optimized, but the differences to the final
routes should be minimal.

Signed-off-by: Jason Lynch <jason@calindora.com>