Build and install rosa as part of the deployment process rather than including the binary.
This finally removes one of the most fragile elements of deployment and
the thing most likely to cause an issue (if minor) for a potential
future person attempting to deploy this web site. However, it does still
rely on the chosen Docker base image remaining available/compatible and
also on the rosa repository remaining available. Testing and updating
the Docker build periodically is highly recommended.

Closes: #1

Signed-off-by: Jason Lynch <jason@aexoden.com>