Add initial all bosses routes.
These are not really intended for general use, as they branch off the
no64-rosa routes. While they are fully optimized, they have not gone
through the second-phase optimization that reduces the number of step
locations. In addition, the encounter timing is largely copied from
other formations and may not be wholly accurate.

Signed-off-by: Jason Lynch <jason@aexoden.com>